Organizing repo structure, adding jupyter notebook files and preprocessed dataset files
