+++ .gitignore
@@ -1,0 +1,2 @@
+# Ignore all zip files
+*.zip
