+++ Original Scripts/Data_loader.py
@@ -1,0 +1,29 @@
+import torch.utils.data as da
+import torch
+
+
+# Get cpu or gpu device for training.
+device = "cuda" if torch.cuda.is_available() else "cpu"
+print("Using {} device".format(device))
+
+
+def data_loader(data, label, batch=64, shuffle=True, drop=False):
+    """
+    Preprocess the data to fit model.
+    Feed data into data_loader.
+    input:
+        data (float): samples*length*ch (samples*ch*length).
+        label (int): samples, ie.: [0, 1, 1, 0, ..., 2].
+        batch (int): batch size
+        shuffle (bool): shuffle data before input into decoder
+        drop (bool): drop the last samples if True
+    output:
+        data loader
+    """
+    label = torch.LongTensor(label.flatten()).to(device)
+    if data.shape[1] >= data.shape[2]:
+        data = torch.tensor(data.swapaxes(1, 2))
+    data = torch.unsqueeze(data, dim=1).type('torch.FloatTensor').to(device)
+    data = da.TensorDataset(data, label)
+    loader = da.DataLoader(dataset=data, batch_size=batch, shuffle=shuffle, drop_last=drop)
+    return loader
