+++ Original Scripts/Train_ANN.py
@@ -1,0 +1,92 @@
+import numpy as np
+import torch
+from torch import nn
+from Data_loader import data_loader
+
+torch.backends.cudnn.benchmark = True
+
+# Get cpu or gpu device for training.
+device = "cuda" if torch.cuda.is_available() else "cpu"
+print("Using {} device".format(device))
+
+
+def initialize_weights(model):
+    for m in model.modules():
+        if isinstance(m, nn.Conv2d):
+            nn.init.kaiming_normal_(m.weight, mode="fan_out", nonlinearity="relu")
+        elif isinstance(m, (nn.BatchNorm2d, nn.GroupNorm)):
+            nn.init.constant_(m.weight, 1)
+            nn.init.constant_(m.bias, 0)
+
+
+def train_ann(ann_model, train_x, train_y, test_x, test_y, ep=500, batch=64):
+    """
+    input:
+        train_x, test_x (float): samples*length*ch (samples*ch*length).
+        train_y, test_y (int): samples, ie.: [0, 1, 1, 0, ..., 2].
+        ep (int): total train and test epoch
+        batch (int): batch size
+    output:
+        train acc, test acc, weight_file
+    """
+    # Define training configuration
+    criterion = nn.CrossEntropyLoss()
+    optimizer = torch.optim.Adam(ann_model.parameters(), lr=0.01)
+    lr_scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=ep)
+
+    # Define data loader
+    train_loader = data_loader(train_x, train_y, batch=batch)
+    test_loader = data_loader(test_x, test_y, batch=batch)
+
+    train_acc = []
+    test_acc = []
+    for epoch in range(ep):
+        # Train ANN
+        ann_model.train()
+        train_loss = 0
+        correct = 0
+        total = 0
+        loss = 0
+        print('\n')
+        for batch_idx, (inputs, targets) in enumerate(train_loader):
+            optimizer.zero_grad()
+            outputs = ann_model(inputs)
+            loss = criterion(outputs, targets)
+            loss.backward()
+            optimizer.step()
+
+            train_loss += loss.item()
+            _, predicted = outputs.max(1)
+            total += targets.size(0)
+            correct += predicted.eq(targets).sum().item()
+            print(batch_idx, len(train_loader), 'Epoch: %d | ANN: trainLoss: %.4f | trainAcc: %.4f%% (%d/%d)'
+                  % (epoch, train_loss / (batch_idx + 1), 100. * correct / total, correct, total))
+
+        lr_scheduler.step()
+        train_acc.append(round(correct / total, 4))
+
+        # Test ANN
+        ann_model.eval()
+        val_loss = 0
+        correct = 0
+        total = 0
+        loss = 0
+        with torch.no_grad():
+            for batch_idx, (inputs, targets) in enumerate(test_loader):
+                outputs = ann_model(inputs)
+                loss = criterion(outputs, targets)
+                val_loss += loss.item()
+                _, predicted = outputs.max(1)
+                total += targets.size(0)
+                correct += predicted.eq(targets).sum().item()
+                print(batch_idx, len(test_loader), 'Epoch: %d | ANN: testLoss: %.4f | testAcc: %.4f%% (%d/%d)'
+                      % (epoch, val_loss / (batch_idx + 1), 100. * correct / total, correct, total))
+
+        test_acc.append(round(correct / total, 4))
+
+    train_acc = np.asarray(train_acc[-1])
+    test_acc = np.asarray(test_acc[-1])
+    return train_acc, test_acc,ann_model
+
+
+
