+++ Original Scripts/model_2a.py
@@ -1,0 +1,221 @@
+import torch
+import torch.nn as nn
+
+
+class LENet(nn.Module):
+    """
+        LENet Model
+    input:
+         data shape as: batch_size*1*channel*length (64*1*22*1000) BCI IV-2a
+         batch_size：64
+         channel：22
+         length：1000
+    output:
+        classes_num
+    """
+
+    def __init__(self, classes_num=4):
+        super(LENet, self).__init__()
+        self.drop_out = 0.5
+
+        self.block_TCB_1 = nn.Sequential(
+            # Temporal Convolution block kernel_size (1,64) #
+            nn.ZeroPad2d((32, 31, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 64),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+        self.block_TCB_2 = nn.Sequential(
+            # Temporal Convolution block kernel_size (1,32) #
+            nn.ZeroPad2d((16, 15, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 32),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+        self.block_TCB_3 = nn.Sequential(
+            # Temporal Convolution block kernel_size (1,16) #
+            nn.ZeroPad2d((8, 7, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 16),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+
+        self.TCB_fusion = nn.Sequential(
+            # Temporal Convolution block fusion kernel_size (1,1) #
+            nn.Conv2d(
+                in_channels=24,
+                out_channels=24,
+                kernel_size=(1, 1),
+                bias=False,
+            ),
+            nn.BatchNorm2d(24)
+        )
+
+        self.SCB = nn.Sequential(
+            # Spatial Convolution block kernel_size (channel,1) #
+            nn.Conv2d(
+                in_channels=24,
+                out_channels=16,
+                kernel_size=(22, 1),
+                groups=8,
+                bias=False
+            ),
+            nn.BatchNorm2d(16),
+            nn.ReLU(),
+            nn.AvgPool2d((1, 4)),
+            nn.Dropout(self.drop_out)
+        )
+
+        self.FFCB = nn.Sequential(
+            # Feature Fusion Convolution block kernel_size (1,16) and (1,1) #
+            nn.ZeroPad2d((7, 8, 0, 0)),
+            nn.Conv2d(
+                in_channels=16,
+                out_channels=16,
+                kernel_size=(1, 16),
+                groups=16,
+                bias=False
+            ),
+            nn.Conv2d(
+                in_channels=16,
+                out_channels=16,
+                kernel_size=(1, 1),
+                bias=False
+            ),  #
+            nn.BatchNorm2d(16),
+            nn.ReLU(),
+            nn.AvgPool2d((1, 8)),
+            nn.Dropout(self.drop_out)
+        )
+
+        self.CCB = nn.Sequential(
+            # Classification Convolution block kernel_size (1,1) #
+            nn.Conv2d(
+                in_channels=16,
+                out_channels=classes_num,
+                kernel_size=(1, 1),
+                bias=False
+            ),
+            nn.AdaptiveAvgPool2d((1, 1)),
+            nn.Flatten(),
+        )
+
+    def forward(self, x):
+        x1 = self.block_TCB_1(x)
+        x2 = self.block_TCB_2(x)
+        x3 = self.block_TCB_3(x)
+        x4 = torch.cat([x1, x2, x3], dim=1)
+        x = self.TCB_fusion(x4)
+        x = self.SCB(x)
+        x = self.FFCB(x)
+        x = self.CCB(x)
+        return x
+
+
+class LENet_FCL(nn.Module):
+    def __init__(self, classes_num=4):
+        super(LENet_FCL, self).__init__()
+        self.drop_out = 0.5
+        self.block_TCB_1 = nn.Sequential(
+            nn.ZeroPad2d((32, 31, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 64),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+        self.block_TCB_2 = nn.Sequential(
+            nn.ZeroPad2d((16, 15, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 32),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+        self.block_TCB_3 = nn.Sequential(
+            nn.ZeroPad2d((8, 7, 0, 0)),
+            nn.Conv2d(
+                in_channels=1,
+                out_channels=8,
+                kernel_size=(1, 16),
+                bias=False,
+            ),
+            nn.BatchNorm2d(8)
+        )
+
+        self.TCB_fusion = nn.Sequential(
+            nn.Conv2d(
+                in_channels=24,
+                out_channels=24,
+                kernel_size=(1, 1),
+                bias=False,
+            ),
+            nn.BatchNorm2d(24)
+        )
+
+        self.SCB = nn.Sequential(
+            nn.Conv2d(
+                in_channels=24,
+                out_channels=16,
+                kernel_size=(22, 1),
+                groups=8,
+                bias=False
+            ),
+            nn.BatchNorm2d(16),
+            nn.ReLU(),
+            nn.AvgPool2d((1, 4)),
+            nn.Dropout(self.drop_out)
+        )
+
+        self.FFCB = nn.Sequential(
+            nn.ZeroPad2d((7, 8, 0, 0)),
+            nn.Conv2d(
+                in_channels=16,
+                out_channels=16,
+                kernel_size=(1, 16),
+                groups=16,
+                bias=False
+            ),
+            nn.Conv2d(
+                in_channels=16,
+                out_channels=16,
+                kernel_size=(1, 1),
+                bias=False
+            ),  #
+            nn.BatchNorm2d(16),
+            nn.ReLU(),
+            nn.AvgPool2d((1, 8)),
+            nn.Dropout(self.drop_out)
+        )
+
+        self.FCL = nn.Sequential(
+            nn.Flatten(),
+            nn.Linear(16 * 31, classes_num)
+        )
+
+    def forward(self, x):
+        x1 = self.block_TCB_1(x)
+        x2 = self.block_TCB_2(x)
+        x3 = self.block_TCB_3(x)
+        x4 = torch.cat([x1, x2, x3], dim=1)
+        x = self.TCB_fusion(x4)
+        x = self.SCB(x)
+        x = self.FFCB(x)
+        x = self.FCL(x)
+        return x
